Warning fix
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@31079 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/import/3ds2lev/3ds2lev.cpp
@@ -356,7 +356,7 @@ int main (int argc, char * argv[])
         // get the numbers in the current mesh
         csFPrintf(stderr, "===================================================\n");
         csFPrintf(stderr, "%-14s  %5ld  %5ld  %5d    %s\n",
-	    mesh->name, mesh->faces, mesh->points, -1, " ");
+	    mesh->name, (long)mesh->faces, (long)mesh->points, -1, " ");
     }
 
     exit(0);
